Make the modelVerifier function-finding simpler
Summary:
I had a pretty hard time understanding what was going on here, the
underlying logic that could be very clear and type-safe was being
obsucred: we were trying to find a global function, then if we didn't
find one getting a local one, but instead of nesting the control flow
we were producing two different optional values and then pattern matching
the pair.

This commit uses control flow to make it explicit that we're looking up
a global function, then trying a local function if we don't find the global.

I wouldn't say it's easy to read now, the actual logic is still pretty fiddly,
but at least the types reflect what's actually going on.

Reviewed By: arthaud

Differential Revision: D63498402

fbshipit-source-id: db60d0bbd1e9cf2eedfa031dde71f82e2cd630b2