Decouple modelTest from SourceCodeApi
Summary:
When I moved all Pysa access to Pyre environments into
PyrePysaEnvironment I punted on a few tests, but we really should
clean those up.

This one was accessing a TypeEnvironment and a SourceCodeApi, both of
which are implementation details of Pyre and shouldn't be leaked. We
already have a `source_of_qualifier` helper in the pyre pysa environment,
so let's use it.

The test code actually gets simpler as a bonus.

Reviewed By: rchen152

Differential Revision: D63547514

fbshipit-source-id: 0cd12bee1918b44c223624543851c1fae9c8a59e