Decouple `TestHelper` from direct use of `Analysis`
Summary:
This required some refactoring, because Pysa does want to collect
analysis errors as a sanity check of the test sources, and that
makes sense.

The solution was to expose `AnalysisError` inside of a new module
`PyrePysaLogic.Testing` so that it's clearly namespaced as being
exposed only for testing uses, not for prod logic.

Reviewed By: arthaud

Differential Revision: D63547516

fbshipit-source-id: 59199907926f7a4a86f0aa03eb93830aeea34d2e