Add user-declared breadcrumbs as local features of the first frame
Summary:
# Problem
Breadcrumbs directly attached to user-declared taint (for instance, using `TaintSource[Test, Via[foo]]`) appear in the issue but do NOT appear on any frame. This is because those are stored as **propagated** breadcrumbs on the "declaration" frame, which is not actually exported to SAPP. Only the "origin" frame (i.e first frame) is exported.

See also the test D63631358.

# Solution
We need custom logic to mark those **propagated** breadcrumbs of "declaration" frames as local breadcrumbs on the first frame (i.e the "Origin" frame). This can be done easily during the "apply_call" logic.

Reviewed By: tianhan0

Differential Revision: D63632529

fbshipit-source-id: a56471a008b7cef18543bd92379ed1cd7eb8b878