allow `get` with unknown key on TypedDict
Summary:
Requested in https://fb.workplace.com/groups/pyreqa/permalink/8523050397784717/

The typing spec isn't clear on what behavior is expected in this case:

https://www.internalfb.com/code/fbsource/[92cbed7884abb76b43e205b8a7260626710d895e]/fbcode/tools/pyre/facebook/conformance/third_party/typeddicts_operations.py?lines=43-44

If we don't know about the key we can't really check it and we'll have to infer `object | None`. I think Pyright infers `Any | None` in this case.

Reviewed By: stroxler

Differential Revision: D63540819

fbshipit-source-id: 1789e9fbea6e9aee9da60d585c2f1f1089e57bf6