Expose queries to the variance map
Summary:
The plumming efforts are not yet done. We need to expose queries, such as get_class_summary to the variable map.

After this, we are able to work on the attached unit test as a starting point.

Note:  I was not sure how to expose the variable_map in the typeOrderTest, so I copied the implementation. This is not ideal, but variable_map requires globalResolution. It's unclear how we can get that in this test. Happy to get any suggestions on this point!

Reviewed By: stroxler

Differential Revision: D63615975

fbshipit-source-id: 779f56b3572828c023cfa278c2000915c5b48aaa