Rename get_successors -> get_parents
Summary:
This callback, which returns the immediate parents of a class, is used
in the computation of successors which is the full MRO.

It's confusing to have it misnamed, because (a) the type doesn't really tell
you whether you are getting just immediate parents or all successors, and (b)
it's plausible that the algorithm would use recursion (it does not).

Renaming the callback fixes all these readability issues.

Reviewed By: rchen152

Differential Revision: D63587244

fbshipit-source-id: acc38a94c0b59067d25c122f99814ab8b19c73ec