Simplify the C3 linearization a bit
Summary:
The use of a function to find the head was confusing because it caused us
to inject an unnecessary name `linearizations` that was really just an
alias to the top-level `linearized_successors` input.

This commit gets rid of that, which simplifies understanding how the data
flows through the algorithm. I also renamed the head-stripped version
to eliminate name shadowing.

Reviewed By: rchen152

Differential Revision: D63647137

fbshipit-source-id: 1fc1442e1efc30b5771f02d47475ec4acafac35c