Add a cycle_detection parameter to the variance_map
Summary:
We need `cycle_detection` so that we can parse annotations. For example, to get the base classes, we find them in the class summary and use `parse_annotation` to convert them to a type, so we can extract the type parameters and their variance.

I don't think we should expose `cycle_detection` to other files, which is why the mli file of attributeResolution is unmodified.

I initialize `cycle_detection` to the empty cycle delection when invoking the `variance_map` method.

Wondering if this is the right way of doing it, hence RFC.

Reviewed By: stroxler

Differential Revision: D63658565

fbshipit-source-id: ad2e769977a73f8185662a5f2e2b1505bef49e5c