Rename things / improve comment on `complex_order`
Summary:
I renamed `order` to `complex_order` for lack of a better name, because
the name shadowing is pretty confusing if we name this global variable `order`.

I renamed the classes to have valid class names A-E instead of numbers 0-4,
the numbers were really confusing to me and can't be used if we were to try
to power this test off of a code snippet in python.

The comment diagram seemed wrong, there was a `b` sitting there that had no
actual node in the test data, and there was an extraneous edge from `bottom`
to the node now named `D`.

Reviewed By: grievejia

Differential Revision: D63663632

fbshipit-source-id: dcd725de069963159fec68c9b79655893d9d71ff