Adjust comments and names in butterfly test
Summary:
I'm finding the MRO tests hard to understand right now. They are often
using *numbers* as class names, which works but doesn't map well to my
brain since you can't actually write class names like that.

In addition, some of the comments are slightly unhelpful - in particular,
the butterfly comment was written upside down.

This diff adjusts the butterfly test by making it clear which way is "up"
in the comment with a diagram, and using A-D instead of 0-3 for the class
names.

Reviewed By: grievejia

Differential Revision: D63661939

fbshipit-source-id: 89c524a085682a0bd5681e75be1b5a0dfe2376ec