Fix a ParamSpec unit test
Summary:
I mistakenly thought that I got an error on callable types like this error:

'Invalid type [31]: Expression `typing.Callable[(L, M)]` is not a valid type.'

This is not true. What's happening is that the program typechecks if we express the type like this:

`Invalid type [31]: Expression `typing.Callable[([L], M)]` is not a valid type.`. So we replaced `L` with `[L]`.

It's unclear yet why this is happening but this should be fixed.

Reviewed By: yangdanny97

Differential Revision: D63842256

fbshipit-source-id: 4c689f0b28b72612740f41e2528a8ae520cf94d4