Add fixme threshold to pyre-upgrade fixme subcommand
Summary:
When we do migrations and/or upgrades there are files where we introduce numerous type errors.

The resulting diffs are often hard to land and can feel "noisy" to end developers. This adds the option to set a threshold for fixmes (similar to what already exists for upgrade-to-strict) and flip the mode to ignore all if too many errors are in a specific python file.

Differential Revision: D63998429

fbshipit-source-id: 0e73aef9a2c20da268186cdfa2d3785736c7f944