Add a bivariant option to post inference variance
Summary:
We need to set our "undefined" variances to bivariance at the beginning of our algorithm. So we add bivariance to post inference variance.

I am not yet sure how we will deal with it in constraint resolution. There are two places where we need to, which I marked by a TODO.

We don't have to infer bivariance, but we do have to start with it when we convert pre inference to post inference at the start of our algorithm.

Reviewed By: samwgoldman

Differential Revision: D64014279

fbshipit-source-id: c95b6c6fffef8440bcdacccee626e464edcd4bce