Add test cases illustrating a refinement bug
Summary:
We recently had a bug report about refinement on a final attribute
mysteriously not working, and it turns out that this is somehow tied to
an equality check on a sub-attribute.

This commit records the problem in two tests the highlight the strangely
specific nature of the bug: only the equality check seems to break refinement,
not even a function call in the same position replicates the behavior (which also
pretty clearly indicates that it isn't a temporary refinement issue.)

Also, rather strangely, the refinement on the outermost variable `b` no longer seems
to be marked as a refinement but as a bare type. This is also unexpected, although
doesn't really impact user-facing behaviors.

Reviewed By: grievejia

Differential Revision: D64146512

fbshipit-source-id: 92c499a8e7985e81942d65f012a24a126795c26e