Improve and rename `Target.override_to_method`
Summary:
There are two types of usage:
1. Get the define name, by first converting `Override` to `Method` and then
getting the define name from `Method`.
2. Get the issue handle name of an `Override` (see D34832215).

Hence, in this diff:
1. We define new APIs in `Target.Regular` for Usage 1.
2. We rename `override_to_method` into `override_for_issue_handle` to describe
why converting an `Override` to `Method` -- we want to get an issue handle that
does not change, when adding a subclass to a base class that previously has no
subclass -- this would otherwise change the issue handle from `foo` to
`Ovr{foo}`.

This diff also renames `Target.define_name` into `Target.define_name_exn` to
explicate that this function might raise exceptions.

Reviewed By: arthaud

Differential Revision: D64258273

fbshipit-source-id: d0378f76729c9ff7e60005b2ca735744a7d25dba