Test isolation dir again to see if it helps with type check durations in the wild
Summary:
This diff adds an isolation dir back in when users pass a GK. The preemptible flag is also removed when a separate isolation dir is used, since we want to see if we can speed up type checks with this approach.

There is a possible issue where a user saves many times and queues up a bunch of sequential type checks, but I think this should be alright since all queued checks should run at the same time after Buck stops blocking.

Let's see if this is an improvement or not.

Reviewed By: MaggieMoss

Differential Revision: D64331815

fbshipit-source-id: a31efb8c5f92b00b2e324cae547e16c6a23d5452