emit invalid annotation error if RHS of type alias fails to parse
Summary:
when we try to parse the RHS of a type alias assignment, we should emit an error if the parsed result is `Top` since that means the thing we were parsing isn't a type.

see https://fb.workplace.com/groups/pyreqa/posts/8681539681935787/?comment_id=8681600518596370&reply_comment_id=8681912865231802 for context.

Credit to tejasreddyvepala for discovering this bug

Reviewed By: stroxler

Differential Revision: D64416189

fbshipit-source-id: 2b3ffa4d3a743b4ea4496c0fc997b6759e59669e