Enable sharding in the IDE
Summary: Similar idea to D64331815. We want to test out sharding in the IDE, but we should make sure we don't absolutely trash UX around it. This lets us double check that sharding works and actually speeds up type checking. If it looks good, then we'll enable it by default.

Reviewed By: grievejia

Differential Revision: D64418245

fbshipit-source-id: 91abe3f1fd5a4a5632cd3c4bfc58547673ecccc4