Use PyrePysaLogic.Testing in Pysa unit tests
Summary:
Several of the Pysa unit tests reach into Pyre internals. It's not
worth trying to eliminate this (changedPathsTest in particular is testing
chnangedPaths, which is inhrently pretty exposed to the low-level logic
of ModuleTracker) but for good hygiene it's nice to thread them through
PyrePysaLogic.Testing so we can see exactly what Pysa relies on.

Reviewed By: yangdanny97

Differential Revision: D64538838

fbshipit-source-id: aee5c1872fee7ac6e34e528a4f6820c8c767478d