Add test cases illustrating dynamic variables in class bodies
Summary:
A recent change to qualification began qualifying statement-level variable
introductions in class bodies more consistently.

Unfortunately this has created a problem where we generate errors whenever
a user binds a variable in a statement that isn't analyzed as part of the
ClassSummary logic, e.g. a `for` or a `while` statement.

The problem doesn't seem to extend to the walrus operator, which is expression-scope;
I suspect we wind up qualifying it as a local but for the moment the important
thing is that if we want to solve the problem we can focus on `while` and `for`.

Reviewed By: rchen152

Differential Revision: D64552737

fbshipit-source-id: 6043fff789b158da098568edfe187da7f5dd0404