Always qualify `with` and `for` targets as locals
Summary:
In D61954297, we changed `for` and `with` targets to qualify as if they were
class attributes.

This makes some sense in that it aligns with runtime behavior (they do in fact
define class attributes, because the runtime takes all names defined at the end
of the class body and turns them into attributes), but it *doesn't* align with
our ClassSummary logic, and it causes very confusing user-facing errors.

Unfortunately, trying to align ClassSummary wouldn't work well because it's not
really practical to infer the types of these attributes and there's no easy
way to make `TypeCheck.ml` treat them as flow-sensitive locals rather than as
class attributes, so we'd get typing holes.

In existing cases where people use these, they are almost always functioning
as local variables - for example, a for-loop that populates a data structure
that is a class attribute. In that context the fact that the for loop binding
becomes an attribute is probably an accident rather than the actual user
intent.

As a result, I think the best user experience we can produce without a lot of
effort is to go back to treating these as locals of the toplevel, and tracking
them in `Resolution.ml` but not in `ClassSummary` or any downstream environment
before `typeCheck`.

Reviewed By: grievejia

Differential Revision: D64563868

fbshipit-source-id: 256326bc23b5e06f205cdc1f708fc691edf6d310