Resolve `Name.Identifier`s into non-object targets
Summary:
So that later the higher order call graph building can find out which targets
may an identifier evaluate into. Currently, an identifier can only resolve into
`global_targets` or `non_local_targets` in `IdentifierCallees`, which are all
`Object` targets. This diff allows resolving an identifier into non-object
targets, which serve as the targets that will be propagated or used to
construct `Target.Parameterized` in higher order call graph building.

Since such targets are "internal" -- they are only used by the higher order
call graph building, `IdentifierCallees.all_targets` do not include such
targets.

**Perofrmance Regression.** The new targets in `callGraphTest.ml` are because
for each `Expression.Call`, we now additionally resolve the targets for the
callee. For example, at call site `foo()`, we now additoinally resolve the
targets for identifier `foo`.  This is unnecessary and hence not ideal. To
avoid this, we may want to update `Ast.Visit.MakeNodeVisitor.visit_expression`
so that when visiting an `Expression.Identifier`, we can avoid visit this
expression when its parent expression is `Expression.Call`, by for example
remembering each expression's parent expression when calling
`visit_expression`. This diff does not implement this, but we can do this in a
follow-up when necessary.

**What is next.** A later diff will similarly resolve `Name.Attribute` into
non-object targets.  For example, assume `test.py` is
```
def bar(): ...
def foo():
  return bar  # test.bar
```
Then we want to resolve the returned expression `test.bar` into function
`test.bar`.

Reviewed By: arthaud

Differential Revision: D64777975

fbshipit-source-id: 40160132ceba6122b35f529f60adc1ed3fd99aee