Support infer variance
Summary:
We begin support for `infer_variance` and add a unittest.

TODO:
Three conformance tests that respectively use:
- dataclass
- property
- final

I looked at is setting the "self" parameter to be bivariant, which seems correct according to Vitousek's writeup on the variance for self types https://fb.workplace.com/notes/flow-inference-redesign/variance-of-the-this-parameter-in-methods/2804686066457515 , but doing so is not giving me the right results.

Reviewed By: stroxler

Differential Revision: D64702381

fbshipit-source-id: 53db8adaea91b46f4b8891465dbcecfbc1c30dac