Disable variance on recursive types
Summary:
As per the latest discussions on recursive types, pyre1 currently does not support generic recursive types.

In legacy syntax, to express a recursive type, we need to define a recursive type alias. But recursive type aliases do not properly support generic types as we do not support that type aliases have type parameters in their representation.

This is unlike for example classes, where the representation itself supports that classes have type parameters.

To begin addressing recursive types, it seems like a good first step is to allow type aliases to have type parameters, so we can query for them like we do for classes.

This is Steven's stack for doing so for classes D61685056. It is a non trivial effort, but looking there seems like a good starting point.

Reviewed By: samwgoldman

Differential Revision: D64990551

fbshipit-source-id: 41e606c02e0a4d9899341ab12f546e7166b78de1