Do not use infer_variance with covariant and contravariant flags
Summary:
We should not use infer_variance when we've already defined covariance or contravariance flags.

We could not catch this before because the variable definitions are not expressive enough.

In this diff, we fix this issue by extending variable declarations to include the infer_variance flag, as existing variable declarations erased that information. Then when we reach typecheck, we can check the flag, along with the variance information we parsed and raise the right error message.

This diff includes adding the new error message defintion and a unit test.

We can now see the conformance improvement.

Reviewed By: stroxler

Differential Revision: D65041792

fbshipit-source-id: 4b5a1e391b348f8fb786140fd3882aa1289a7a8c