Retry preempted type check only if type check was interrupted by another type check
Summary:
This diff adds a guard around Buck subprocess calls to record the number of concurrent operations. Since async operations can't concurrently modify the set, if the set is not empty, then there's another type check operation going on.

We can use this to figure out if we should retry a preempted Buck operation. If the list is empty, and we were preempted, then we need to retry, since no other type checks are in process. Otherwise, there's another type check in progress, and we most likely preempted ourselves.

The big benefit to doing this is in speeding up type checks. Type checks with the preemptible flag have terrible performance, since we're counting from the start of getting type errors to when we either get type errors or give up. This may take multiple rounds of being preempted, so if we can avoid counting the extra long durations for those, we can get a better estimate of how much time we're actually spending getting type errors.

asyncfixfire

Reviewed By: grievejia

Differential Revision: D64936276

fbshipit-source-id: 96e22cdf5d4e8b7c7a0c8fdc7c7d896116f164d1