Add GK check to see if we should allow preempting
Summary:
The previous diff added logic around figuring out if we should retry a type check if it was preempted.

With that logic present, this diff adds logic to check a gatekeeper and see if we should allow the preemptible flag to be passed to the subprocess call.

Along with that, we need to make sure any time we type check in the default isolation directory, we **always** use the preemptible flag. We should never block user buck operations because of an IDE type check. This should be handled by the GK,  but there's also going to be an extra check here to make sure we don't accidentally mess this up and cause issues.

Reviewed By: grievejia

Differential Revision: D64979502

fbshipit-source-id: 600e6b510740f6d65a48fcd004f5628a0c2b246e