Prevent users from shadowing `builtins.pyi` in ptt
Summary:
When using buck-based per-target type checking, we realised that
it is possible for users, by setting `base_module = ''` and having
an `__init__.py`, to shadow `builtins.py`.

This is pretty confusing because there's no user-visible indication
that this is happening, and we just get tons of type errors on all
the builtin types being undefined.

Reviewed By: grievejia

Differential Revision: D65076559

fbshipit-source-id: d1ea05c3d0b5c6a8b2e114ba42fb7b81e2f1f1a8