Add the ability to query attributes on a generic class
Summary:
Prior to this commit, when Pyre encountered an `attributes(...)` query on
a generic class it would raise an IncorrectParameters exception, which
would then give an error
```
`"Type `Whatever`has the wrong number of parameters"
```

The logic to raise this error came, I think, from the `LessOrEqual` query
where we presumably wanted to fuss at anyone trying to do type order on
generic types that lack type arguments.

But for the Attributes query, which returns *uninstantiated* attributes coming
directly off of the class definition, this isn't reasonable: we should definitely
allow a generic class to be used.

Reviewed By: vladima

Differential Revision: D65170764

fbshipit-source-id: d74ccbac37946e1d6c2574f2bc60f6a87d19d417