Add a failing test case for a regression in no-deamon-query
Summary:
We have a regression in no-daemon-query where a type check that hits
WeakenMutableLiterals on an unknown type can crash a query, which is
affecting the glean indexer. This is impacting the glean indexer.

Before I fix the problem, I want to be sure we have a test case demonstrating
the problem, so that we don't re-introduce a similar regression (recent work
improving type information for a transpiler partner team led to the issue).

This commit adds a failing test, the next one will fix it.

Reviewed By: kinto0

Differential Revision: D65164187

fbshipit-source-id: f0cc9b78cc7922ea1272989138ceca030c589220