Recover from `ClassHierarchy.Untracked` in `compute_local_annotations`
Summary:
In D62449018 we modified the way LocationBasedLookup.ml gets type information to directly
take types coming out of a type check. This allowed us to fix problems affecting a partner
team where expressions that create scope (like comprehensions) weren't analyzed successfully,
but it also introduced a regression to the no-daemon-query mode we set up for Glean.

The problem is that we hacked up Pyre incoherently to support Glean, by making some parts of
Pyre allow untracked classes, while other parts of Pyre still have no idea how to handle the
use of unknown types. Because of this, in no-daemon-query mode it is fairly common for type
check to crash if it hits some branch that isn't robust. Prior to D62449018 this would cause
only local gaps, because the exception was caught in TypeAliasEnvironment and we woudl just
pull empty state out of shared memory. But with D62449018 the exception is triggered directly
and crashes the entire query, which just returns an error.

(Why are we no longer using shared memory? Because we can't really afford to store the new
expression-level types data in shared memory, it would balloon saved states of the IG codebase
too much. For this reason, we just compute them as needed with a fresh type check.)

This commit reverts the error handling to be consistent with the earlier behavior, by wrapping
the direct invocation of type check in a `try / with` block and returning `None` in the optional
result.

Reviewed By: kinto0

Differential Revision: D65164188

fbshipit-source-id: 2f60c045e25ef6944faa607fc9a572c762192aa7