Use `as_identifier_reference` to resolve `Name.Attribute`
Summary:
So that we can improve the performance -- calling `resolve_regular_callees` is
expensive, as we see a 100% perf regression during the call graph step in D64867871.

For a `Name.Attribute`, when no function target can be resolved by calling
`as_identifier_reference` (e.g., `x = obj.foo` where `obj` is a local variable
or a parameter),  technically we could call `resolve_callees_from_type` to find
the callees, but in this diff we decided to not do this for better performance.

Reviewed By: arthaud

Differential Revision: D65084402

fbshipit-source-id: 2f2d3c688f911c6b804dcaf0d49d0a00aeb85473