Add deriving show to lots of `typeCheck.ml` data
Summary:
I have been trying to drill into a perf regression, and we have quite a few
non-printable types. Adding show to them (they are all trivially derivable)
to enable faster debugging, might as well commit the change.

This does require a local suppression of the unused-code warning: I'm fairly
sure this is why the show was missing, because we didn't want to allow unused
code but an unused printer is flagged. I think the local suppression is the right solution -
we still don't want general unused code, but it's silly to have to go in and tag a bunch of
datatypes before you can start print debugging.

Reviewed By: yangdanny97

Differential Revision: D65280662

fbshipit-source-id: df4731564478ae2f9f612b871c1fc7dd51533365