Rename `instantiate_recursive_type_with_solve`
Summary:
This function name had `protocol` all over the place, but it actually is
shared between both of the recursive structural types in Pyre: protocols
and recursive aliases both use it.

This diff is just fixing up the name, as well as the name of the `<:`
callback, to better reflect this (which also makes it easier to see,
when debugging, when we are in protocol-specific vs general recursive
type code).

Reviewed By: rchen152

Differential Revision: D65303118

fbshipit-source-id: e5dc0391c74a2aaea511454e6198fd2c8ae3d47d