Rename `sanitized_candidate`
Summary:
The fact that we transform `candidate` to mark formerly free vars as bound
and shadow the variable name is not exactly unreasonable, but given the size of
the function it's a little hard to track when we ware looking at the original
vs sanitized candidate. This is especially true since we have a lack of comments
explaining how all this works, so skimming the code is our only option for trying
to understand it.

Renaming makes it a bit easier to follow.

Reviewed By: rchen152

Differential Revision: D65303120

fbshipit-source-id: 8f7b56fb438e34963cbb12f9931b43af043c35a2