Regroup logic in `instantiate_protocol_parameters_with_solve`
Summary:
This code is at the heart of a large slowdown in numpy-related type checking
due to some interaction of union types with recursive protocols.

I need to make it easier to navigate the code; given how Pyre currently nests
things, it is generally easier to understand if we always group data used only
by one part of the functionality of a large function right next to the use
case (which limits how long each variable is in scope, and makes it easier to
see which bits of functionality are related).

This commit is just making that kind of change in
`instantiate_protocol_parameters_with_solve` - pushing several high-level
variables down into branches if they aren't needed in multiple places.

Reviewed By: grievejia

Differential Revision: D65303124

fbshipit-source-id: f056f891eaf461b487c6d208f4fe67235c82e31c