Treat formal arguments and local variables as the same variant under `TaintAccessPath.Root` in `HigherOrderCallGraph.State`
Summary:
This is because the expressions and statements in the control flow grpahs use
the qualified names for both formal arguments and local variables.  Currently,
we use unqualified names for formal arguments, and use qualified names for
local variables, although conceptually formal arguments and local variables are
the same from the perspective of propagating the function-typed values that are
associated with them.

This diff treats formal arguments and local variables uniformly, by using the
same variant (under `TaintAccessPath.Root`) to represent both.
Note that in `Target.Parameterized`, we still keep the variant obtained from
calling `TaintAccessPath.match_actuals_to_formals arguments`. This is useful to
distinguish different types of parameters. However, when creating a
`HigherOrderCallGraph.State` from a `Target.Parameterized`, we convert the
parameter into `TaintAccessPath.Root.Variable`.

Reviewed By: arthaud

Differential Revision: D65248733

fbshipit-source-id: af587954fcf6172ebcb38cef3a7f43448654765f