Close over the `order` in solver recursion
Summary:
In order to fix a perf issue in handling recursive aliases, I want to
add a dynamic-programming-style cache of recursive type instantiations to the
constraint solver.

It is much easier to do this at the code level if we already are closing over
compute context in the recursive calls (e.g. `solve_less_or_equal` calling
itself) instead of manually passing context.

I actually think this closing style is a little clearer anyway becausee
otherwise as a reader of the code you have to read hundreds of invocations in
order to be certain that we aren't doing somethign crazy like redefining the
`order` somewhere - this style makes it obvious right at the top of the
function that `order`

Reviewed By: yangdanny97

Differential Revision: D65398921

fbshipit-source-id: f84dacd780086eccda8c9534c7b993afcc83db87