Tweak the signature of instantiate-recursive-type funcitons
Summary:
I want to be able to close over `order` more reasonably, which will make
it much easier to add a dynamic-programming-style cache for expensive
recursive type solving in Pyre.

A first step to make this work is to make the signature of
`instantiate_protocol_parameters`, which has an optional labeled
argument, friendlier to closing over order. I just make `candidate` a
positional argument.

Since the function isn't really public (it's in the interface for testing
only, in production it's only used by other parts of ConstraintsSet) the
stripping of the label is not a significant interface change.

Reviewed By: yangdanny97

Differential Revision: D65398916

fbshipit-source-id: e8a609149c73869cd6d0fc61a4f609db4b61b89c