Add a test demonstraing slow analysis
Summary:
We've observed problems dealing with numpy's `ArrayLike` protocol in Pyre.

We tracked the problem to recursive protocol solving: it turns out that the
complexity of solving a recursive type when the protocol applies to a container
that can be nested is essentially `O( a ^ n )` where
- `a` is the number o attributes that branch (e.g. methods taking Self as input),
  and includes sub-terms for any unions of protocols involved (which happens for
  the `ArrayLike` example) and overloads of functions (which is relevent for the
  arraylike example)
- `n` is the amount of nesting of container types involved: at each level we fan
  out attribute checks, so this factor is exponential.

The new test passes, but if you set `dump_time` to `true` you can see that it
is pretty slow; on my laptop it takes 14 seconds to run.

Reviewed By: yangdanny97

Differential Revision: D65398919

fbshipit-source-id: 7324235578c55a110a1bab4f9f72d0aac2bdcea8