Add a cache of recursive instantiations
Summary:
# What problem are we solving, and how?

This commit solves a severe perf problem Pyre ran into on the numpy
ArrayLike protocol, which was caused by the fact that without memoization
solving a nested type that might be a recursive prodocol is `O( a * n )`
where
- `a` is the cost of solving the attribute table and has subterms for number
  of attributes involving the recursive type as well as overloads and unions
  of protocols
- `n` is the depth of the nested container type in question

The complexity drops to nothing major if we memoize the recursive solve; the
perf issue is due to a huge compute tree where the leaves are overwhelmingly
redundant.

This is essentially the dynamic programming pattern: take a tree of calculations that is
impractical to traverse directly but has massive redundancy and memoize.

# Correctness: why memoization is safe here

This commit is a little tricky to reason about because in principle memoizing
could cause problems if it interacted with other impure parts of Pyre. I
am fairly confident it does not because:
- We account for `cycle_detections` in the memoization key (this is important,
  I'm fairly sure we would introduce bugs in edge cases if we did not).
- We are only caching recursive type instantiation, and that does not
  accept a `constraints` argument. So there is no risk of solve calls under
  different constraints polluting one another.
- The cache is 100% internal to the mutual recursion of a single solve, and
  `ConstraintsSet` has no knowledge of environment tables. If we ever fan
  out from solving to an environment table (which can happen, e.g. when
  getting attributes of protocols), the cache is dropped at the point when
  we hit an environment layer. So there's no risk that this cache would
  mess up incremental graph structures, it is only capable of caching over
  calculations that had the same dependency key (in which case skipping
  redundant fetches has no impact on the graph).

# Performance results on the arraylike example added in D65398919

In a branch of the code with significant logging, I saw a smaller version of
this example go from more than 80K `solve_less_or_equal` calls and more than
0.95 seconds of compute time to fewer than 3K `solve_less_or_equal` calls and
less than 0.08 seconds of compute time (and that 0.8 was inflated by all
the logging I was doing).

On this slightly larger example comput time drops from more than 13 seconds
to less than 0.6 seconds.

Reviewed By: grievejia

Differential Revision: D65398918

fbshipit-source-id: 872db8dbd02d95eec09397af7a97e38bd75e3451