Rename `builtins_type` to `class_type`
Summary:
At some point I renamed `Type.meta`, which is the function to construct ocaml
data representing an annotation like `type[C]`, to `type.builtins_type`.

Jia noted at the time that the new name was potentially confusing because
one might think `builtins_type int` means "the built-in-type int" whereas
it really means `builtins.type[int]`.

I had trouble thinking of a better name for a while, but I realized recently
that the real point of `type[...]` is that it has to be a valid class
object (or a `GenericAlias` that wraps a class object that has been
specialized) in order to have the required runtime behaviors.

So I think `class_type` is a decent name for it, probably an improvement
over both `meta` and `builtins_type`.

Happy to entertain any ideas reviewrs think might be even better.

Reviewed By: grievejia

Differential Revision: D65461323

fbshipit-source-id: d44161c369c308b1fabcae9111303396280cdc9d